% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admission.R
\docType{data}
\name{admission}
\alias{admission}
\title{MBA admissions data.}
\source{
The data were taken from Johnson and Wichern (2008).
}
\usage{
data(admission)
}
\value{
A data frame with 85 rows and 3 variables. The data frame contains
the following columns:
\describe{
\item{GPA}{The applicant's grade point average (GPA) on a 0.0 - 4.0 scale.}
\item{GMAT}{The applicant's graduate management admission test (GMAT) score
on a 200 - 800 scale.}
\item{decision}{A factor with three levels, \code{admit}, \code{border} and
\code{notadmit}, which refer to the category to which the student was assigned by
admissions tutors (admit, borderline or do not admit).}}
}
\description{
Data on applicants to the Masters of Business Administration (MBA) programme of a US
business graduate school.
}
\examples{
data(admission)
head(admission)
}
\references{
Johnson, R.A and Wichern, D.W. (2008) \emph{Applied Multivariate Statistical}
\emph{Analysis, Sixth Edition.} Pearson.
}
\keyword{datasets}
