% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durhamSLR.R
\docType{package}
\name{durhamSLR}
\alias{durhamSLR}
\title{The durhamSLR package}
\description{
Data for Statistical Learning modules at Durham University.
}
\author{
Sarah Heaps \email{sarah.e.heaps@durham.ac.uk}
}
