% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heptathlon.R
\docType{data}
\name{heptathlon}
\alias{heptathlon}
\title{Heptathlon data.}
\source{
The data were taken from Wikipedia at
\url{https://en.wikipedia.org/wiki/Athletics_at_the_2012_Summer_Olympics_\%E2\%80\%93_Women\%27s_heptathlon}.
}
\usage{
data(heptathlon)
}
\value{
A data frame with 29 rows and 7 variables. The data frame contains
the following columns:
\describe{
\item{H100M}{100 metres hurdles (seconds).}
\item{HJ}{High jump (metres).}
\item{SP}{Shot put (metres).}
\item{R200M}{200 metres (seconds).}
\item{LJ}{Long jump (metres).}
\item{JT}{Javelin throw (metres).}
\item{R800M}{800 metres (seconds).}}
}
\description{
Results for the heptathlon at the 2012 Olympic Games in London for the 29 athletes who
completed all events and were not disqualified.
}
\examples{
data(heptathlon)
head(heptathlon)
}
\keyword{datasets}
